node-main/deps/v8/test/inspector/debugger/set-breakpoint-on-instrumentation-expected.txt

Test if breakpoints are hit that are set on instrumentation pause in js

Running test: testSetBreakpointOnInstrumentationPause
set breakpoint and evaluate script..
Setting breakpoint at instrumentation break location
Paused at foo.js with reason "instrumentation".
Hit breakpoints: []
Paused at foo.js with reason "other".
Hit breakpoints: ["4:0:20:3"]
Done.

Running test: testSetConditionalBreakpointTrueConditionOnInstrumentationPause
set breakpoint and evaluate script..
Setting breakpoint at instrumentation break location
Paused at foo.js with reason "instrumentation".
Hit breakpoints: []
Paused at foo.js with reason "other".
Hit breakpoints: ["4:0:20:4"]
Done.

Running test: testSetConditionalBreakpointFalseConditionOnInstrumentationPause
set breakpoint and evaluate script..
Setting breakpoint at instrumentation break location
Paused at foo.js with reason "instrumentation".
Hit breakpoints: []
Done.
